+++ src/core/datastore/sync.js
@@ -42,6 +42,11 @@ export class SyncManager {
     this.collection = collection;
 
     /**
+     * @type {string}
+     */
+    this.tag = options.tag || undefined;
+
+    /**
      * @type {Client}
      */
     this.client = options.client || Client.sharedInstance();
@@ -80,7 +85,8 @@ export class SyncManager {
       query: query,
       properties: options.properties,
       timeout: options.timeout,
-      client: this.client
+      client: this.client,
+      tag: this.tag
     });
     return request.execute()
       .then(response => response.data)
@@ -102,7 +108,8 @@ export class SyncManager {
           properties: options.properties,
           query: syncQuery,
           timeout: options.timeout,
-          client: this.client
+          client: this.client,
+          tag: this.tag
         });
         return request.execute()
           .then(response => response.data);
@@ -172,7 +179,8 @@ export class SyncManager {
         }),
         properties: options.properties,
         query: query,
-        timeout: options.timeout
+        timeout: options.timeout,
+        tag: this.tag
       });
       return findRequest.execute()
         .then(response => response.data)
@@ -195,7 +203,8 @@ export class SyncManager {
             }),
             properties: options.properties,
             body: syncEntity,
-            timeout: options.timeout
+            timeout: options.timeout,
+            tag: this.tag
           });
           return request.execute();
         });
@@ -244,7 +253,8 @@ export class SyncManager {
           properties: options.properties,
           query: query,
           timeout: options.timeout,
-          client: this.client
+          client: this.client,
+          tag: this.tag
         };
         let request = new KinveyRequest(config);
 
@@ -309,7 +319,8 @@ export class SyncManager {
                     }),
                     properties: options.properties,
                     timeout: options.timeout,
-                    client: this.client
+                    client: this.client,
+                    tag: this.tag
                   });
                   return request.execute()
                     .then(() => {
@@ -322,7 +333,8 @@ export class SyncManager {
                           pathname: `${this.pathname}/${syncEntity._id}`
                         }),
                         properties: options.properties,
-                        timeout: options.timeout
+                        timeout: options.timeout,
+                        tag: this.tag
                       });
                       return request.execute();
                     })
@@ -352,7 +364,8 @@ export class SyncManager {
                       pathname: `${this.backendPathname}/${entityId}`
                     }),
                     properties: options.properties,
-                    timeout: options.timeout
+                    timeout: options.timeout,
+                    tag: this.tag
                   });
                   return request.execute()
                     .then(response => response.data)
@@ -400,7 +413,8 @@ export class SyncManager {
                               pathname: `${this.pathname}/${syncEntity._id}`
                             }),
                             properties: options.properties,
-                            timeout: options.timeout
+                            timeout: options.timeout,
+                            tag: this.tag
                           });
                           return request.execute()
                             .then(() => {
@@ -414,7 +428,8 @@ export class SyncManager {
                                 }),
                                 properties: options.properties,
                                 timeout: options.timeout,
-                                body: entity
+                                body: entity,
+                                tag: this.tag
                               });
                               return request.execute()
                                 .then(response => response.data);
@@ -430,7 +445,8 @@ export class SyncManager {
                                     pathname: `${this.backendPathname}/${entityId}`
                                   }),
                                   properties: options.properties,
-                                  timeout: options.timeout
+                                  timeout: options.timeout,
+                                  tag: this.tag
                                 });
 
                                 return request.execute()
@@ -536,7 +552,8 @@ export class SyncManager {
               pathname: `${this.pathname}/${entity._id}`
             }),
             properties: options.properties,
-            timeout: options.timeout
+            timeout: options.timeout,
+            tag: this.tag
           });
           return request.execute()
             .then(response => response.data);