+++ src/core/index.js
@@ -1,7 +1,9 @@
+import './polyfills';
 export { client, getAppVersion, setAppVersion, init, initialize, ping } from './kinvey';
 export { Acl } from './acl';
 export { Aggregation } from './aggregation';
 export { AuthorizationGrant } from './identity';
+export { Client } from './client';
 export { CustomEndpoint } from './endpoint';
 export { DataStore, DataStoreType, SyncOperation } from './datastore';
 export { LiveService } from './live';
@@ -9,7 +11,7 @@ export { Files } from './files';
 export { Log } from './log';
 export { Metadata } from './metadata';
 export { Query } from './query';
-export { Properties } from './request';
+export { Properties, StorageProvider } from './request';
 export { User } from './user';
 export {
   ActiveUserError,