+++ src/html5/index.js
@@ -1,1 +1,2 @@
+export { StorageProvider } from './storage';
 export * from './kinvey';