+++ src/nativescript/cache/index.ts
@@ -1,82 +1,41 @@
 import { isDefined, isEmpty } from '../../core/utils';
-import { Middleware, RequestMethod, StatusCode, Storage as CoreStorage } from '../../core/request';
+import { CacheMiddleware as CoreCacheMiddleware, Storage as CoreStorage, StorageProvider as CoreStorageProvider } from '../../core/request';
 import { sqLite } from './sqlite';
+
+export const StorageProvider = Object.assign({}, CoreStorageProvider, {
+  SQLite: 'SQLite'
+});
+Object.freeze(StorageProvider);
 
 class Storage extends CoreStorage {
   name: string;
+  storageProviders: Array<string>;
 
-  constructor(name: string) {
-    super(name);
+  constructor(name, storageProviders = [StorageProvider.SQLite, StorageProvider.Memory]) {
+    super(name, storageProviders);
   }
 
   loadAdapter() {
-    return sqLite.load(this.name)
-      .then((adapter) => {
-        if (!isDefined(adapter)) {
-          return super.loadAdapter();
+    return this.storageProviders.reduce((promise, storageProvider) => {
+      return promise.then((adapter) => {
+        if (adapter) {
+          return adapter;
         }
 
-        return adapter;
+        switch (storageProvider) {
+          case StorageProvider.SQLite:
+            return sqLite.load(this.name);
+          default:
+            return super.loadAdapter();
+        }
       });
+    }, Promise.resolve());
   }
 }
 
-export class CacheMiddleware extends Middleware {
-  constructor(name = 'NativeScript Cache Middleware') {
-    super(name);
-  }
-
-  handle(request: any) {
-    const {
-      method,
-      body,
-      appKey,
-      collection,
-      entityId,
-      encryptionKey
-    } = request;
-    const storage = new Storage(appKey);
-    let promise;
-
-    if (method === RequestMethod.GET) {
-      if (entityId) {
-        promise = storage.findById(collection, entityId);
-      } else {
-        promise = storage.find(collection);
-      }
-    } else if (method === RequestMethod.POST || method === RequestMethod.PUT) {
-      if (entityId === '_group') {
-        promise = storage.find(collection);
-      } else {
-        promise = storage.save(collection, body);
-      }
-    } else if (method === RequestMethod.DELETE) {
-      if (!collection) {
-        promise = storage.clear();
-      } else {
-        promise = storage.removeById(collection, entityId);
-      }
-    }
-
-    return promise
-      .then((data) => {
-        const response = {
-          statusCode: method === 'POST' ? StatusCode.Created : StatusCode.Ok,
-          data: data
-        };
-
-        if (method === 'POST' && entityId === '_group') {
-          response.statusCode = StatusCode.Ok;
-        }
-
-        if (!data || isEmpty(data)) {
-          response.statusCode = StatusCode.Empty;
-        }
-
-        return response;
-      })
-      .catch(error => ({ statusCode: error.code }))
-      .then(response => ({ response: response }));
+export class CacheMiddleware extends CoreCacheMiddleware {
+  loadStorage(name, storageProviders) {
+    return new Storage(name, storageProviders);
   }
 }